backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.25
python-dotenv>=1.0
pydantic>=2.6
anthropic>=0.40
reportlab>=4.0
playwright>=1.40
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.2